move imported dependencies to lib folder
