+++ lib/feedWatcher.js
